Q1Q3UTILITY/requirements.txt
pillow==11.0.0
easygui==0.98.3
vgio==1.3.0
